refactor(minesweeper): minesweeper 패키지 추가 및 domain 패키지 생성
도메인 관련 객체들은 전부 domain 패키지로 분리
