+++ src/main/kotlin/minesweeper/MineSweeperGame.kt
@@ -1,11 +1,11 @@
-package ui
+package minesweeper
 
-import GameBoard
-import ui.ConsoleInput.inputColumnLength
-import ui.ConsoleInput.inputCountOfLandmine
-import ui.ConsoleInput.inputRowLength
-import ui.ConsoleOutput.announceGameStarted
-import ui.ConsoleOutput.displayCurrentGameBoard
+import minesweeper.domain.GameBoard
+import minesweeper.ui.ConsoleInput.inputColumnLength
+import minesweeper.ui.ConsoleInput.inputCountOfLandmine
+import minesweeper.ui.ConsoleInput.inputRowLength
+import minesweeper.ui.ConsoleOutput.announceGameStarted
+import minesweeper.ui.ConsoleOutput.displayCurrentGameBoard
 
 fun main() {
     val rowLength = inputRowLength()
