+++ src/main/kotlin/minesweeper/domain/RandomShuffleAlgorithm.kt
@@ -1,3 +1,5 @@
+package minesweeper.domain
+
 class RandomShuffleAlgorithm : ShuffleAlgorithm {
     override fun <T> shuffle(list: List<T>): List<T> = list.shuffled()
 }
