+++ src/main/kotlin/minesweeper/domain/ShuffleAlgorithm.kt
@@ -1,3 +1,5 @@
+package minesweeper.domain
+
 interface ShuffleAlgorithm {
     fun <T> shuffle(list: List<T>): List<T>
 }
