+++ src/main/kotlin/minesweeper/domain/Vulture.kt
@@ -1,3 +1,5 @@
+package minesweeper.domain
+
 class Vulture {
     fun plantMine(cell: Cell): Landmine = Landmine(row = cell.row(), column = cell.column())
 }
