+++ src/main/kotlin/minesweeper/ui/ConsoleInput.kt
@@ -1,4 +1,4 @@
-package ui
+package minesweeper.ui
 
 object ConsoleInput {
     fun inputRowLength(): Int {
