+++ src/test/kotlin/minesweeper/domain/BasicCellGridTextFixture.kt
@@ -1,3 +1,5 @@
+package minesweeper.domain
+
 object BasicCellGridTextFixture {
     val threeByThreeGrid: List<List<Cell>> = List(3) { row -> List(3) { column -> BasicCell(row = row + 1, column = column + 1) } }
 }
