+++ src/test/kotlin/minesweeper/domain/BasicCellTest.kt
@@ -1,8 +1,10 @@
+package minesweeper.domain
+
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 
 class BasicCellTest : BehaviorSpec({
-    given("BasicCell 은") {
+    given("domain.BasicCell 은") {
         val sut = BasicCell(row = 1, column = 1)
 
         `when`("기본 상태의 display 값은") {
