+++ src/test/kotlin/minesweeper/domain/FixedShuffleAlgorithm.kt
@@ -1,3 +1,5 @@
+package minesweeper.domain
+
 class FixedShuffleAlgorithm : ShuffleAlgorithm {
     override fun <T> shuffle(list: List<T>): List<T> = list
 }
