+++ src/test/kotlin/minesweeper/domain/GameBoardTest.kt
@@ -1,3 +1,5 @@
+package minesweeper.domain
+
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.collections.shouldHaveSize
