+++ src/test/kotlin/minesweeper/domain/LandmineTest.kt
@@ -1,3 +1,5 @@
+package minesweeper.domain
+
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
