+++ src/test/kotlin/minesweeper/domain/RowTest.kt
@@ -1,3 +1,5 @@
+package minesweeper.domain
+
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
